*  Slide TH1I under the hood of CSpectrum1DL. *  Make daqdatatypes.h use RTypes.h for e.g. UInt_t *  Resolve the fallout of Root type decisions (e.g. Size_t is a real of    some sort).

+++ main/Core/BitSpectrumL.cpp
@@ -337,7 +337,7 @@ CBitSpectrumL::CreateStorage()
   setStorageType(keLong);
 
   Size_t        nBytes   = StorageNeeded();
-  UInt_t*      pStorage = new UInt_t[nBytes/sizeof(UInt_t)];
+  UInt_t*      pStorage = new UInt_t[static_cast<unsigned>(nBytes/sizeof(UInt_t))];
 
   ReplaceStorage(pStorage);	// Storage now owned by parent.
   Clear();
